requirements.txt
Babytracker==0.0
Chameleon==2.7.1
Mako==0.5.0
MarkupSafe==0.15
PasteDeploy==1.5.0
Pygments==1.4
SQLAlchemy==0.7.4
WebOb==1.2b2
pyramid==1.3a3
pyramid-debugtoolbar==0.9.7
pyramid-tm==0.3
repoze.lru==0.4
transaction==1.2.0
translationstring==0.4
venusian==1.0a2
wsgiref==0.1.2
zope.deprecation==3.5.0
zope.interface==3.8.0
zope.sqlalchemy==0.7
